Organized routes controllers files. 2 new routes: /api/stock/:symbol and /api/stock/add/:symbol.
